refactor(scheduler): consolidate time-domain resource lookup and reduce index storage footprint
- Change PUSCH/PUCCH time-domain resource indices from `unsigned` to `uint8_t` (bounded by MAX_NOF_PUSCH_TD_RES_ALLOCS=16) across dl_time_domain_mapper and ul_time_domain_mapper
- Rename accessor methods to clarify they return indices: pdsch_td_resources()/pdsch_td_res_indices() and pusch_td_resources()/pusch_td_res_indices()
- Simplify internal per-slot storage from full resource copies to index-based lists, reducing memory usage
- Move pusch_td_resource_indices.cpp from ocudu_sched to sched_config CMake target and wire ul_time_domain_mapper constructor to call real get_pusch_td_resource_indices_per_slot()
- Simplify API: get_pusch_td_resource_indices_per_slot() now accepts pusch_td_list span and min_k1 directly instead of pusch_cfg_common/dl_data_to_ul_ack/ss_info
- Replace ocudu_sanity_check() with proper error handling: log warning and skip gracefully when no valid PDCCH slot exists for a UL slot
- Simplify k1 candidate configuration: use pucch_auto_resources with min_k1 instead of pre-computing k1 candidates via pucch_explicit_resources

+++ lib/scheduler/support/pucch/pucch_k1_helper.cpp
@@ -12,7 +12,7 @@ std::vector<static_vector<uint8_t, 8>> ocudu::get_pucch_k1_list_per_slot(
     span<const uint8_t>                                       dl_data_to_ul_ack,
     const std::optional<tdd_ul_dl_config_common>&             tdd_cfg_common,
     const std::vector<pusch_time_domain_resource_allocation>& pusch_td_alloc_list,
-    const std::vector<static_vector<unsigned, pusch_constants::MAX_NOF_PUSCH_TD_RES_ALLOCS>>&
+    const std::vector<static_vector<uint8_t, pusch_constants::MAX_NOF_PUSCH_TD_RES_ALLOCS>>&
         pusch_td_resource_indices_per_slot)
 {
   ocudu_assert(not dl_data_to_ul_ack.empty(), "dl_data_to_ul_ack cannot be empty");
